pyproject.toml
[project]
name = "drf-p33-channels"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "channels-redis>=4.3.0",
    "channels[daphne]>=4.3.2",
    "django>=5.2.8",
    "django-ckeditor-5>=0.2.18",
    "django-cors-headers>=4.9.0",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.11",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-django>=4.11.1",
    "python-dotenv>=1.2.1",
    "ujson>=5.11.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "root.settings"
# -- recommended but optional:
python_files = ["test_*.py", "*_test.py", "testing/python/*.py"]
asyncio_mode = "auto"